Update useServers to remove unneeded destructured elements
